server2/requirements.txt
Django==2.2
djangorestframework==3.9.2
mysqlclient==1.4.2.post1
celery==4.3.0
django-environ==0.4.5